# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        },
        {
            "url": "/"
        }
    ],
    "paths": {
        "/api/otp": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Send OTP to email",
                "description": "Sends an OTP to the specified email address for verification",
                "operationId": "0f777b671cb868d22a159c9364fd43e1",
                "requestBody": {
                    "description": "Email address",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP sent successfully"
                    },
                    "400": {
                        "description": "Invalid email"
                    }
                }
            }
        },
        "/api/verify-otp": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Verify OTP code",
                "description": "Verifies the OTP code sent to the specified email",
                "operationId": "9c58f8dd50999f6e4ddd20104f6e1174",
                "requestBody": {
                    "description": "OTP and email address",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "otp"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "otp": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP verified successfully"
                    },
                    "400": {
                        "description": "Invalid OTP or email"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}